Remove interfaces from xconfmap
