+++ confmap/confmap.go
@@ -49,3 +49,25 @@ type Unmarshaler = internal.Unmarshaler
 // A configuration struct can implement this interface to override the default
 // marshaling.
 type Marshaler = internal.Marshaler
+
+// ScalarValue provides access to a scalar configuration value and allows
+// calling back into the confmap decoding/encoding machinery.
+//
+// This interface is only provided to methods used for [ScalarUnmarshaler] and
+// [ScalarMarshaler] implementations and cannot be implemented by types outside
+// the confmap package.
+type ScalarValue = internal.ScalarValue
+
+// ScalarUnmarshaler is an interface which may be implemented by wrapper types
+// to customize their behavior when the type under the wrapper is a scalar
+// value.
+//
+// This should be used for types like `Wrapper[T]` where T is a scalar type, and
+// the wrapper type needs to implement custom logic for unmarshaling from a
+// scalar value (e.g. `5` for `Wrapper[int]`) into the wrapper type (e.g.
+// `Wrapper[int]{inner: 5}`).
+type ScalarUnmarshaler = internal.ScalarUnmarshaler
+
+// ScalarMarshaler is an interface which may be implemented by wrapper types
+// to customize their behavior when the type under the wrapper is a scalar value.
+type ScalarMarshaler = internal.ScalarMarshaler
