+++ confmap/internal/testdata/scalar.yaml
